CBI: Allow board to compose data from other sources
Currently, the source of CBI data is only EEPROM. This patch allows
CBI data to be composed of other sources such as ADC and a chip register.

cbi_board_override is called by CBI library when a data is being
requested. Boards can implement this callback to add additional bits
to the data.

A board itself may need to get CBI data first to manipulate data
properly. For such a case, a board can return EC_ERROR_BUSY to indicate
to inform the callers that the data is not fully ready while a board
itself can accept EC_ERROR_BUSY as expected.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/129569858
BRANCH=none
TEST=Read LCM_ID properly

Change-Id: Ie1f962c64c8d1461a6c171bc6c6d0c855c82e945