factory_mode: refactor factory_enable_failed
Refactor factory_enable_failed, so cr50 always resets if a reset is
requested. This change also renames factory_enable_failed to be more
specific. It renames ccd_hook_active to wait_for_factory_ccd_change so
it's obvious what the variable is doing. It's waiting for the ccd_config
change after we enable factory mode.

Enabling factory mode can fail in a lot of ways, but by the time we
called factory_enable_failed, the failure is specifically about saving
the config. This change renames the function, so the failure is a bit
more specific.

If a reset is required, always reset the system even if saving the
factory config failed. ccd_reset_factory_failed is triggered if the ccd
changed hook isn't triggered quickly enough or if cr50 fails to save the
ccd config. Cr50 has already wiped the TPM and has most likely saved
some if not all of the factory mode state. Cr50 should still reset even
if the config isn't saved to be safe.

enable_ccd_factory_mode isn't used in the process to enable factory mode
during init, so this change won't cause a cr50 reboot loop from cr50
trying and failing to enable factory mode during init. This only affects
the RMA and factory mode enable vendor commands.

BUG=b:129956462
BRANCH=cr50
TEST=Use rma and factory mode vendor commands to enable factory mode.

Change-Id: Ib8a502297040296fb0a2250a9e8945af330d4334
Signed-off-by: Mary Ruthven <mruthven@chromium.org>