Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)
> 
> Also, are you able to remove the corresponding usage in unhandled_vector, as well as the procedure definition of get_current_interrupt_vector? That would be ideal.

No, we cannot remove it.

We can only remove it from the ISR which has vector information in compile time. While defining a normal ISR, we specify vector number which is used for EOI to IOAPIC. so a normal ISR has the vector information and we can remove "get_current_interrupt_vector" from it

But unhandled_vector is the ISR for all other vectors which should detect which vector is called it.

Patch-set: 2
Reviewer: Gerrit User 1137654 <1137654@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>