Casta: Increase board_hibernate() wait to avoid hard resets
When hibernating with AC plugged in, casta will try negotiating down to
5V.  This negotiation may sometimes take longer than the currently
allocated sleep, causing the port partner to go into error recovery
(re-sending PS_RDY, followed by soft resets, followed by a hard reset).
This hard reset will wake the system back up from hibernation.

Increasing this sleep to 300 ms appears to give enough time for a
GoodCRC to be send to the PS_RDY message from the charger, avoiding the
hard reset.

Note that hibernating on AC is not a customer or factory requirement, as
the typical EC hibernate scenario is when there is no external power and
the battery needs to be conserved.

BRANCH=octopus
BUG=b:130687403
TEST=flashed to casta EVT, ran hibernate key sequence several times with
AC plugged in and verified system hibernated

Change-Id: Ia40ee63f3cf9a244ba35b4cc700e2b41eea1dde5
Signed-off-by: Diana Z <dzigterman@chromium.org>