motion sense: Calculate loop time based on sensor needs
Currently the motion sense loop bases its sleep time based on the
fastest active sensor. This method has several flaws, first it does not
take into account any task switching overhead and second with a mix of
interrupt driven and forced sensors the sleep time gets recalculated
every time there is an interrupt causing the loop to oversleep, third if
multiple sensors do not have rates that are in sync the timing of the
slower sensor will be off. For example if there was a sensor running at
50 Hz and one running at 20 Hz the slower sensor would end up being
sampled at about 16 Hz instead of 20 Hz.

This change calculates an ideal read time for every forced mode sensor
and calculates the sleep time based on the nearest read time. Every time
a sensor is read the next read time is calculated based on the ideal read
time not the actual read time so that reading does not drift because of
system load or other overhead.

BUG=b:129159505
TEST=Ran sensor CTS tests on arcada, without this change the
     magnetometer was failing 50 Hz tests at about 38 Hz with 30% jitter
     with this change in place 50 Hz was spot on with about 10% jitter
BRANCH=none

Change-Id: Ia4fccb083713b490518d45e7398eb3be3b957eae
Signed-off-by: Mathew King <mathewk@chromium.org>