grunt: Improve safe state when entering DP mode
When configuring DisplayPort alt mode, follow the spec more closely:
Place the USB Type-C pins that are to be re-configured to DisplayPort
Configuration into the Safe state before sending the configure command.
Then switch the pins to DisplayPort after receiving the ack to the
command.

For TYPEC_MUX_DOCK, the superspeed signals can remain connected. For
TYPEC_MUX_DP, disconnect the superspeed signals in svdm_dp_config,
then re-configure the pins to DisplayPort in svdm_dp_post_config.

This means we avoid an unnecessary disconnection in the TYPEC_MUX_DOCK
case (CL:1553572) but still follow the spec and put the pins in safe
state in the TYPEC_MUX_DP case.

BUG=b:123310411
BRANCH=grunt
TEST=External display works with both DOCK and DP pin modes.

Change-Id: I7de990e7dae053d089027cdc62094e5f8cd5ec4b
Signed-off-by: Edward Hill <ecgh@chromium.org>