bq25710: Adjust settings to limit VSYS droop with no battery
When there is no battery present VSYS is set to the value defined in
register 0x3E. The reset value of this register is based on what type
of battery pack is used. In addition to this, big core systems such as
CML can a peak power draw that exceeds the external USB PD adapters
limit. This can lead to a VSYS collapse which then will cause the AP
to sequence out of S0.

This CL modifies 3 default settings for the bq25710 to reduce the
likelihood of VSYS droop leading to the AP exiting S0. The primary
change here is enabling prochot to be active when VSYS is operating
within a threshold of VSYS_MIN. As mentioned above, in the no battery
case, this will always be true which results in prochot being
continuosly asserted.

In addition to the settings change, the reset register control bit is
set at the beginning of the init routine to ensure that following EC
resets the bq25710 always starts in the same state as it does
following a power cycle event.

BUG=b:130209125
BRANCH=none
TEST=Verfied that when using both a 45W and 60W USB PD external
charger that Hatch can boot successfully without a battery. In
particular tested that without setting BQ25710_PROCHOT_PROFILE_VSYS
hatch fails to boot successfully to the kernel, but does when this bit
is set. Also with scope verified that prochot is continuosly asserted
when there is no battery present.

Change-Id: I450f735bca861baa5aae45af70fd8344a618f609
Signed-off-by: Scott Collyer <scollyer@google.com>