eve USBPD: Don't set CHECK_IDENTITY on startup
This removes a fix for chromium:644663 (CL:415698).

Consequently, CL:1185495 needs to be reverted because it was created
to fix a double entry bug which is caused by setting CHECK_IDENTITY
on startup.

---------- >8 ------------
Currently, TCPM tries to exit alternate mode on shutdown and sets
CHECK_IDENTITY flag on startup.

This behavior originates from CL:415698 (chromium:644663).

When a dual-role monitor is plugged during S5, the TCPM performs
full communication: negotiates power, swaps data role (to become DFP),
discovers identity, then enters the DP mode.

On startup, CHECK_IDENTITY flag is set. This causes the TCPM to send
additional DiscIdentity request (followed by an additional EnterMode).

CL:963844 fixes this issue for Fizz. We can apply it for Nami. It sets
the flag on startup only if the port is UFP. (This is probably
ineffective because UFP ports are not supposed to request DiscIdentity.)

However, since the DP mode is exited during S0->S5 transition, the
subsequent boot fails to enter the DP mode. This is because the port is
DFP and CL:963844 makes the TCPM skip setting the flag.

We can further tweak the logic so that the TCPM sets the flag even for
DFP if the DP mode is not active on boot (or mux is not set to DP).

However, making the logic more and more complex has caveats.

First, some monitors do not (legitimately) implement ExitMode. So, we
can't assume the monitor always exits DP mode as we expect.

Second, there are timing issues. If a power button is pressed first then
a monitor is connected, the flag is set, causing a double mode entry.

So, this patch removes this logic used to set CHECK_IDENTITY flag during
power on and off and instead it disconnects the ports on shutdown.

After disconnect (re-connect), the devices (re-)perform power negotiation,
identity discovery, and mode entry immediately.

BRANCH=firmware-eve-9584.B
BUG=b/130617222,b/110441862,b/75958206
BUG=chromium:887731
TEST=Verify the picture is on the external monitor after:
1. Shutdown the system.
2. Connect DUT to a dual-role USB-C monitor.
3. Boot to OS
TEST=Verify the picture is on the external monitor after:
1. Boot to OS.
2. Connect DUT to a dual-role USB-C monitor.
3. Shutdown the system
4. Boot to OS
TEST=Verify the picture is on the external monitor after:
1. Boot to OS.
2. Connect DUT to a dual-role USB-C monitor.
3. Hardware reset (refresh+power)

Change-Id: Id9bf61b3368fe7a42b035107b87ed9328140dc31
Signed-off-by: Nathan K <nkolluru@google.com>