usbpd: Disconnect ports on shutdown
WIP

Need to fix "TCPC p1 reset!" storm.

Currently, TCPM tries to exit alternate mode on shutdown and sets
DiscoverIdent flag on resume.

When a dual-role monitor is plugged when the system is in S5, the TCPM
performs all: negotiates power, swaps data role (to become DFP),
discover identity, then enters the DP mode.

On startup, CHECK_IDENTITY flag is set. This causes the monitor to be
hit by an aditional DiscIdentity request (followed by an additional
EnterMode).

CL:963844 fixes this issue for Fizz. We can apply it for Nami.

However, since the DP mode is exited during S0->S5 transition, the
subsequent boot fails to enter the DP mode. This is because the port is
DFP and CL:963844 makes the TCPM skip setting the flag.

We can further tweak the logic so that the TCPM sets the flag even for
DFP if the DP mode is not active on boot (or mux is not set to DP).

However, making the logic more and more complex has caveats.

First, some monitors do not (legitimately) implement ExitMode. So, we
can't assume the monitor always exits DP mode as we expect.

Second, there are timing issues. If a power button is pressed first then
a monitor is connected, the flag is set, causing a double mode entry.

So, this patch removes this logic used to set CHECK_IDENTITY flag during
power on and off and instead it disconnects the ports on shutdown.

After disconnect (re-connect), the devices (re-)perform power negotiation,
identity discovery, and mode entry immediately.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/130617222,b/110441862,b/75958206
BUG=chromium:887731
TEST=Verify the picture is on the external monitor after:
1. Shutdown the system.
2. Connect DUT to a dual-role USB-C monitor.
3. Boot to OS
TEST=Verify the picture is on the external monitor after:
1. Boot to OS.
2. Connect DUT to a dual-role USB-C monitor.
3. Shutdown the system
4. Boot to OS
TEST=Verify the picture is on the external monitor after:
1. Boot to OS.
2. Connect DUT to a dual-role USB-C monitor.
3. Hardware reset (refresh+power)

Change-Id: I5ecd55b8747f156d5e5bf29682e969608810d28a