hatch: Account for TCPC reset polarity in ALERT handler
Reset polarity was added to tcpc_config so that the reset function
could be common across different boards which have different TCPC
selections. This was applie to pd_reset_muc(), however, that CL did
not take into account the reset polarity in the tcpc_get_alert_status
function. This CL fixes that oversight.

BUG=b:130194031
BRANCH=none
TEST=verified on Hatch that both ports 0/1 function correctly.

Change-Id: I0d75d3655c799d4c74f4a6fb1805c06c1fe99c06
Signed-off-by: Scott Collyer <scollyer@google.com>