Update patch set 1
Patch Set 1: Verified-1

Code Looks OK.
Just talked with Nick.  There may be a hardware fix coming in the next rev of hardware.
If we are fixing this then leaving it enabled and taking a waiver on the CTS test may be the correct way to go.  We need to come to a meeting of the minds on this.

Patch-set: 1
Reviewer: Gerrit User 1273086 <1273086@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=-1