Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> Allowing swap from source to sink while suspended sounds good to me.
> This seems to be Vincent's question on the original CL:238963.
> Getting stuck as a sink seems fine to me, unless I'm misunderstanding Alec's argument there.

This CL seems reasonable to me. I think switching to a SNK is better for everyone since the situation we are trying to fix is

1) Dock plugged in with no power
2) DUT sleeps
3) Plug in power to Dock

Here is where we should transition to SNK but we don't (this CL changes this behavior)


Let's say we don't submit this CL and continuing the above scenario, if the user unplugged the dock and plugged it back in we would negotiate as a SNK. So the behavior would seem weird to a user without this change.

Patch-set: 1
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>