nvmem: populate default tpm objects after wipeout
TPM memory wipe code is removing all TPM objects from the flash,
including the reserved objects, which are supposed to be always
present. The assumption was that the Cr50 would be reset after the
wipe out and the initialization code would populated the reserved
objects with default values.

But in fact Cr50 reset is not guaranteed after TPM wipeout, so it is
better to call the init function explicitly to make sure that all
reserved objects are in the flash at all times.

BRANCH=cr50, cr50-mp
BUG=b:69907320
TEST='dump' command ran on the Cr50 console after RMA open shows all
     reserved objects present.

Change-Id: Id9e227de0995c6491da9f38fc8ca11df3661c71f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>