ec_commands: Add command to request chips requiring FW update
Add an EC command so that depthcharge can request EC regarding the list
of chips that require FW update in a board. This helps with supporting
multiple chip topologies within a single baseboard.

BUG=b:128820536, b:119046668
BRANCH=None
TEST=Ensure that the device boots to ChromeOS. Ensure that the Aux FW
update information is passed back to depthcharge in response to the new
command.

Bobba:
2019-05-09 14:51:16 [11.727600 System is unlocked. Skip MPU configuration]
2019-05-09 14:51:16 [11.727727 HC 0xab]

Casta:
2019-05-09 15:30:45 [8.792710 System is unlocked. Skip MPU configuration]
2019-05-09 15:30:45 [8.792852 HC 0xab]

Ampton:
2019-05-09 15:38:13 [8.436698 HC 0xab]
2019-05-09 15:38:13 [8.437236 HC 0xab err 1]

Change-Id: I11127c547fa9f2ad6f3c4f242b223591d654260a
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>