ec_commands: Add command to request chips requiring FW update
Add an EC command so that depthcharge can request EC regarding the list
of chips that require FW update in a board. This helps with supporting
multiple chip topologies within a single baseboard.

BUG=b:128820536, b:119046668
BRANCH=None
TEST=Ensure that the device boots to ChromeOS. Ensure that the Aux FW
update information is passed back to depthcharge in response to the new
command.

Casta:
2019-05-01 12:43:08 [8.449000 Jumping to image RW]
...
2019-05-01 12:43:08 [8.510506 System is unlocked. Skip MPU configuration]
2019-05-01 12:43:08 [8.510622 HC 0xab]

Ampton:
2019-05-01 12:40:02 [8.526164 Jumping to image RW]
...
2019-05-01 12:40:03 [8.614505 HC 0xab]
2019-05-01 12:40:03 [8.614798 Error -2 getting aux fw chip list]

Change-Id: I11127c547fa9f2ad6f3c4f242b223591d654260a
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>