cbi: add cbi_check host command
We want to be able to add new EEPROMs to the AVL for CBI, and one of the
requirements is that the EEPROM protects the entire region of data that
is within CBI.

Add a host command that we can call from the host to verify that the
EEPROM contents are indeed read-only when the hardware write protect pin
is asserted on the EEPROM.

BRANCH=none
BUG=b:119065537
TEST=manually tested on phaser that WP work, then made code think that
WP was on when it actually wasn't in hardware to verify failure case.
Verified other result cases as well.

Change-Id: I638a49eff405608fbe48b5a2a21d6e87c1e1762c
Signed-off-by: Jett Rink <jettrink@chromium.org>