Flapjack: Disable charging from BC 1.2 charger as USB-C charger
Currently, if a charger shows Rp=USB on USB-C port, the charge manager
chooses it and sets the max current to 500 mA even if it can provide
higher power as a BC 1.2 charger.

This patch introduces CONFIG_USBC_DISABLE_CHARGE_FROM_RP_DEF. When it's
defined, a BC 1.2 charger won't be recognized as a USB-C charger.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/131353444
BRANCH=none
TEST=Charge Flapjack from BC 1.2 charger on USB-C port.

Change-Id: I50969973026185dd2aecdb768985cd116c1d32f7