Update patch set 3
Patch Set 3:

Can we add this when we need it? So far we have worked around the need for this (or have explicitly not used long division algorithms since we were in a performance critical section)

How does this change the image size for Arcada?

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>