Update patch set 11
Patch Set 11:

Looking at this reminded me of nocturne CL:1286299 which I had meant to investigate bringing over to master but then forgot about :(
Are they both solving the same or similar problem? At first glance the interaction between them is not clear to me. As Jett said, this is complicated!

Patch-set: 11
CC: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>