Update patch set 7
Patch Set 7:

OK, verified the scenario.

condition: event "a"(deadline "A") is already in comparator(here, the actual value in the comparator has 800us more at least)

1. New event "b" is less than event "a" but the diff is extremely small(or zero). event "b" triggers s/w IRQ with deadline "B".
2. ISR(triggered by s/w) clears status and runs process_timers().
3. But before the ISR sets new value to the comparator, timer 1 expires(main counter > deadline "A"+a) for event "a" on deadline A(+a) and set interrupt status bit. This means both event "a" and event "b" are expired so corresponding tasks will get notice.
4. the process_timers() sets comparator for the next event which is the closest.

After the ISR, we get extra interrupt but it does nothing but sets the comparator with the same closest event.

This is what happening now. The extra interrupt doesn't harm system. but if we want avoid this, we need to move the status clearing after process_timers()

Patch-set: 7
Reviewer: Gerrit User 1137654 <1137654@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>