Update patch set 1
Patch Set 1:

How sensitive is the timing for needing to set this register?  If it could be done at the end of the init function, it would probably be better to implement a generic board_init field similar to usb_mux which could be used to re-set the register later.

Patch-set: 1
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>