ec-devutils: Fix fdpexpect import
As of pexpect 3.0, the fdexpect module was moved into the pexpect
package.  Prior to 4.0, a wrapper module was installed along with
pexpect, so "import fdexpect" continued to work.  In 4.0 this wrapper
was removed.

Importing fdexpect from pexpect works for our current pexpect 3.3, and
will continue to work when we update pexpect to 4.0+.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BRANCH=none
BUG=chromium:958677
TEST=(0) with pexpect 3.3, (1) sudo servod --board=grunt
TEST=(0) Uprev pexpect to 4.2.1, (1) sudo servod --board=grunt
 => servod works in both cases

Change-Id: I9d21042713d0cc1826cf55da4af7a16f44b746bb