driver: lsm6dsm: sensor event timing
BUG=b:129159505
BRANCH=None
TEST=Ran Android CTS

This change does 3 things:
1. Fixes event out of order errors in CTS.
2. Fixes missing event errors in CTS.

#1 is fixed by doing the spreading on the ec. A new queue is
used (data_queue) and in place of the old call to
motion_sense_fifo_add_data we now add the event to the queue
and increment the sensor's sample count. At the caller
(load_fifo) we then figure out the window (time between the
last interrupt and the read) and the period by dividing the
window by the number of samples (this is done per sensor).
If the period is larger than the odr, then the odr is used
(this helps with accuracy). Events are now spread between the
known time the first entry was added and the read time.

#2 is done by updating the last_fifo_read_ts and checking
the interrupt GPIO. If we find that the GPIO is still low
at the end of the read that means that we've gotten new data
while reading and never fully emptied the FIFO. We know this
must have happened some time between the time we read the
count and when we were reading the FIFO, for now we'll use
the upper bound of this to be safe.

* Doing this involved aliasing the GPIO for the lsm6dms
  interrupt.

Change-Id: Ie9d4309540238b270e0109bef8da8210482d0588
Signed-off-by: Yuval Peress <peress@chromium.org>