Update patch set 7
Patch Set 7:

> Patch Set 7:
> To my knowledge, the safe but also efficient solution is adding wait WIP after process_timer() at the end of isr as suggested by jett.

This seems like a reasonable solution to me. If you implement this, either summarize this conversation in a comment with the code, or at a minimum link to this discussion so that it will make sense to someone reading in the future.

Patch-set: 7
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>