Adding gpio for max32660 chip.
This is a "preview" patch created from a subset of files in a ZIP archive we received from Maxim.
The plan is to not submit this one, but use it to calibrate my reviewing of the future patches from Maxim, as well as gauge whether there is some high-level feedback we want to give to Maxim, before they break up their lump of code into patches.

One piece of high-level feedback I want us to agree on, is whether we want to insist that they do not include any xxx_api.{h,c} files, copied straight from their own support libraries, adding "glue" code such as what is found in gpio_chip.c here.

Change-Id: Ib5d486caaa5979ebf325e05d928178049476ded9