it83xx/uart: uart2 clock belongs to the EC side
On eSPI platform, we need to set that uart clock belongs to the EC side.
So EC won't get stuck at uart initialization.

This change also disconnect PD port1 if the CONFIG_UART_HOST
option is enabled. Because they are both alternate function on the
same pins(GPIOH1/2).

BUG=none
BRANCH=none
TEST=Enable the CONFIG_UART_HOST option and ensure EC without stuck
     during uart initialization.

Change-Id: I83cc5213cdecfcbae9556ecf0a2a066f8168ccab
Signed-off-by: Dino Li <Dino.Li@ite.com.tw>