Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (11 comments)
> > 
> > These are my first sketch of what I want to have Maxim do with their new EC code.  Please let me know if I am too pedantic, or if I there are other things that I should have pointed out or demanded changed.
> 
> Why don't we ask them to submit a patch for review?

We are going to. Maxim engineers are in the process of setting up gerrit access, and breaking their code up into patches.

I should have made it clear in the subject: This review is not meant to be submitted, this is meant as a way to calibrate my my reviewing.

Currently all we have is a big ZIP file, and I would like to take the opportunity to see if there are any high-level input we want to give Maxim, before they break their files into 20 patches.

Patch-set: 3
Reviewer: Gerrit User 1341996 <1341996@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>