pd_protocol: Add ready_state_holdoff_timer.
Recently a sink holdoff timer was added to the PD stack which allowed
the state machine to prevent initiating any messages for 200ms after
entering the SNK_READY state.  This was to give time for some chatty
sources to send messages to avoid a collision.  Apparently, the same
thing can happen when we are a source (collision with chatty sink).

This commit reuses the holdoff timer for resolution as a source as well,
which starts after an explicit contract is established.

In order to prevent any potential new collisions, some jitter based off
of the system timestamp is added to the holdoff timer.

BUG=b:132202148, chromium:925618
BRANCH=firmware-atlas-11827.B
TEST=Flash atlas, plug in a fully featured C-C cable between atlas and
the LG 27UK850-W, verify that no conflict occurs and external display
always works.
TEST=Verify that no messages are initiated by the source within 200ms of
sending PS_RDY.
TEST=Flash nocturne, verify Dell U3818DW still works over C-C cable.
TEST=Flash nocturne, verify CableMatters MST DP hub still works with
charge through.
TEST=Verify with Twinkie that messages are sent at varied timestamps
between 200-300ms in the SNK/SRC_READY state.

Change-Id: I195199de271950ae09c2b26194ddc5f271b296a0
Signed-off-by: Aseda Aboagye <aaboagye@google.com>