i2c: add virtual port and address scheme
For i2cxfer host command, add virtual port and addresses that will
resolve to the correct physical port and address based on the EC static
configuration.

The first supported virtual address if for CBI's EEPROM.

This change also makes the passthru_allowed callback take the 8-bit i2c
address instead of the 7-bit address because the rest of the EC codebase
works with 8-bit addresses. There was only a single instance.

BRANCH=none
BUG=b:119065537
TEST=Used virtual addresses in new firmware_ECCbiEeprom faft test

Change-Id: I82c2d5462fcb4edbc92ea60765971190fed7ae81
Signed-off-by: Jett Rink <jettrink@chromium.org>