Adds a switch to change the compiler toolchain when building the Sweetberry project to repair ToT
The change which updated the toolchain exposed a problem with the
USB initialization which prevents powerlog from being able to
communicate with the device. The root cause has not been
identified so the follow up issue b/132204142 has been filed.

BRANCH=bnemec/126223732
BUG=b:126223732
TEST=Verified that the Sweetberry project is able to build, lsusb
identifies the device, and powerlog is capable of communicating with
the device and retrieving data.

Change-Id: I0d1f7e79a8ea890dcf5da3c16427220cf9067fdf
Signed-off-by: Brian Nemec <bnemec@google.com>