TEST-ONLY: kukui: Automatically switch to RW after a while
Since we have more goodies in the RW section of the firwmware, it's nice to be running that copy.

Unfortunatelly it's hard to get to RW without a software sync, and updating the software sync EC RW copy during EC development would take way too long. So just hardcode the EC's RO copy to switch to RW automatically.

Note: This is to be used with the software sync flag 0x200 disabled, the EC will automatically switch to the copy it already had (probably matching the RO version) without AP prompting it to.

TEST=Kukui/Krane has "ectool motionsense" working as soon as ec is
flashed by a developer, without the flashed RW copy getting replaced.
BUG=Email thread "Default value for "disable EC software sync" in firmware"
BRANCH=master

Change-Id: If5bc1a6f025099ec31d3272748f2ed175d167394
Signed-off-by: Alexandru M Stan <amstan@chromium.org>