Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> I'm not sure about this. I think we need more discussion. I'm worried it'll be overly used.
> 
> Also #ifdef CONFIG_X... #endif annotates the surrounded code can be removed once CONFIG_X is deprecated. If we can annotate functions/declarations differently, it may work... (e.g. unused_if(CONFIG_X)).

Sure, it is nice to have a little annotation that this function is only used under certain configuration options; however, adding semantical checking of this to the build process essentially boils down to that you need to solve a boolean-satisfiability problem to see if all particular configurations will compile. Even with a trivial codebase, doing this is infeasible, and is part of the reasoning why the Linux kernel adopts a style of adding __maybe_unused, and then notating (in a comment) the particular configuration in which it will be used.

You also have to look at what we already have: by guarding procedure definitions in an ifdef, C source code which only uses functions under a certain configuration now needs to guard every application in an ifdef. In a sense, the IS_ENABLED macro becomes un-usable, and we resort back to deeply nested ifdefs that we have no feasible mechanism to ensure any potential configuration does not cause compile errors.

So, I say we adopt the practice which has been tried and tested on millions of more lines of code than we have (the Linux kernel) with great success. Make sense?

Patch-set: 1
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0