Fleex: Set minimum battery level for Try.SRC and OS boot
Some battery and fuel gague combinations for fleex are unable to
withstand the loss of a charger's Vbus signal when they are reporting 1%
charge.  This change increases the Try.SRC battery required.
Additionally, it bars the OS from booting until we have either PD
negotiated or achieved 2% battery.

BUG=b:128935567
BRANCH=octopus
TEST=builds

Change-Id: Ia0eca111daaf6b63f71c397231de4cfa30e34b8c
Signed-off-by: Diana Z <dzigterman@chromium.org>