flash_log: add api for setting base timestamp
The Cr50 environment does not have a wall clock, which makes it
impossible to associate flash log entries with real time.

This patch provides an API which allows to set a base time value and
then use it plus current Cr50 uptime to generate more sensible flash
log timestamps.

Care is taken to ensure that attempts to set timestamp base such that
it would cause a log timestamps rollback do not succeed.

A unit test is being added to verify this behavior.

BRANCH=none
BUG=b:132287488
TEST='make buildall -j' (which runs the new tests) succeeds.

Change-Id: I7521df1bac5aef67e0cf634c183bf1618655f48d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>