Update patch set 3
Patch Set 3:

(6 comments)

It seems saner to execute the lock separately by EC_CMD_I2C_PASSTHRU_PROTECT. It'll also give a better control and visibility to Depthcharge in terms of what EC (as a binary) is doing and which port is protected, etc.

We have to issue two separate commands per port (lookup & lock). So, the boot time penalty might be non-negligible. Four host commands are executed in 4 ms on Flapjack, FYI.

Patch-set: 3