Update patch set 3
Patch Set 3:

> If that's the plan, we should even more prefer multiple-command approach to single-command approach. Such a generic command will contain data and code used only by particular boards with particular class of chips. Such a command will be complex. The code size will grow. We should push complexity to AP not to EC.

Sorry, I don't understand this reasoning. This is adding a single command that returns a hardcoded array of structs from the data section. That's about at lean as it can possibly get on the EC side. Even if we make the data more complicated, it will still just be a hardcoded struct and the code size impact will just be the data itself.

On the other hand, you're asking for *four* different commands, all with their own command handlers and the code needed to look up information from dynamic structures. And once we add more chip types you'll have to make all those command handlers even more complex, whereas with the other solution it doesn't matter because you're just hardcoding a different struct. I don't see how this could be any more efficient, and on top of that you're roughly doubling the total amount of EC host commands we need to send during a whole normal mode boot. Maybe that's only 20ms, but if we just didn't care every time we wasted 20ms somewhere Chromebooks wouldn't boot as fast as they do.

I think we should design our commands and APIs in a way that is efficient for the normal boot case. If you want to save, I'd try to do that in other areas (e.g. test support... it looks like EC_CMD_I2C_LOOKUP is being added only for that, and what it does could easily be subsumed into this command as just another chip type for the CBI that depthcharge doesn't happen to have an update handler for).

Patch-set: 3
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>